requirements.txt
streamlit 
streamlit-echarts 
Pillow 
